🔥 ファイル削除と、型の整理
