+++ src/features/register/_types/index.ts
@@ -1,0 +1,1 @@
+export type RegisterTab = "gmail" | "api";
