--- src/types/store.ts
@@ -1,2 +1,0 @@
-export type ActiveTab = "home" | "register" | "settings";
-export type RegisterTab = "gmail" | "api";
