+++ src/utils/validation.ts
@@ -1,7 +1,19 @@
-import { z } from "zod";
 import { NewBook, NewSeries } from "@/db/types";
-import { BookValidationResult, SeriesValidationResult } from "@/types/provider.types";
 import { createValidationError } from "@/types/errors";
+import { z } from "zod";
+
+ interface BookValidationResult {
+  isValid: boolean;
+  errors: string[];
+  warnings: string[];
+}
+
+ interface SeriesValidationResult {
+  isValid: boolean;
+  errors: string[];
+  warnings: string[];
+}
+
 
 // ISBN正規表現パターン
 const ISBN_REGEX = /^(?:ISBN(?:-1[03])?:? )?(?=[0-9X]{10}$|(?=(?:[0-9]+[- ]){3})[- 0-9X]{13}$|97[89][0-9]{10}$|(?=(?:[0-9]+[- ]){4})[- 0-9]{17}$)(?:97[89][- ]?)?[0-9]{1,5}[- ]?[0-9]+[- ]?[0-9]+[- ]?[0-9X]$/;
@@ -16,45 +28,45 @@ export const BOOK_VALIDATION_SCHEMA = z.object({
     .min(1, "タイトルは1文字以上である必要があります")
     .max(500, "タイトルは500文字以内である必要があります")
     .transform(str => str.trim().replace(/\s+/g, ' ')),
-  
+
   author: z
     .string({ required_error: "著者名は必須です" })
     .min(1, "著者名は1文字以上である必要があります")
     .max(200, "著者名は200文字以内である必要があります")
     .transform(str => str.trim().replace(/\s+/g, ' ')),
-  
+
   targetUrl: z
     .string({ required_error: "購入URLは必須です" })
     .regex(URL_REGEX, "有効なURLを入力してください"),
-  
+
   isbn: z
     .string()
     .regex(ISBN_REGEX, "正しいISBN形式で入力してください")
     .transform(str => str.replace(/[-\s]/g, ''))
     .optional()
     .or(z.literal("")),
-  
+
   volume: z
     .number()
     .int("巻数は整数である必要があります")
     .min(1, "巻数は1以上である必要があります")
     .max(9999, "巻数は9999以下である必要があります")
     .optional(),
-  
+
   description: z
     .string()
     .max(2000, "説明は2000文字以内である必要があります")
     .transform(str => str?.trim().replace(/\s+/g, ' ') || "")
     .optional(),
-  
+
   imageUrl: z
     .string()
     .url("有効なURL形式で入力してください")
     .optional()
     .or(z.literal("")),
-  
+
   googleBooksId: z.string().optional(),
-  
+
   purchaseDate: z
     .date({ required_error: "購入日は必須です" })
     .refine(
@@ -74,7 +86,7 @@ export const BOOK_VALIDATION_SCHEMA = z.object({
       },
       "購入日が古すぎます"
     ),
-  
+
   seriesId: z.string({ required_error: "シリーズIDは必須です" }),
   shopId: z.string({ required_error: "ショップIDは必須です" })
 });
@@ -85,25 +97,25 @@ export const SERIES_VALIDATION_SCHEMA = z.object({
     .min(1, "シリーズタイトルは1文字以上である必要があります")
     .max(500, "シリーズタイトルは500文字以内である必要があります")
     .transform(str => str.trim().replace(/\s+/g, ' ')),
-  
+
   author: z
     .string({ required_error: "著者名は必須です" })
     .min(1, "著者名は1文字以上である必要があります")
     .max(200, "著者名は200文字以内である必要があります")
     .transform(str => str.trim().replace(/\s+/g, ' ')),
-  
+
   description: z
     .string()
     .max(2000, "説明は2000文字以内である必要があります")
     .transform(str => str?.trim().replace(/\s+/g, ' ') || "")
     .optional(),
-  
+
   thumbnail: z
     .string()
     .url("有効なURL形式で入力してください")
     .optional()
     .or(z.literal("")),
-  
+
   googleBooksSeriesId: z.string().optional()
 });
 
@@ -115,7 +127,7 @@ export type ValidatedSeries = z.infer<typeof SERIES_VALIDATION_SCHEMA>;
 export const validateBook = (bookData: Partial<NewBook>): BookValidationResult => {
   try {
     const result = BOOK_VALIDATION_SCHEMA.safeParse(bookData);
-    
+
     if (result.success) {
       return {
         isValid: true,
@@ -123,10 +135,10 @@ export const validateBook = (bookData: Partial<NewBook>): BookValidationResult =
         warnings: []
       };
     }
-    
+
     const errors: string[] = [];
     const warnings: string[] = [];
-    
+
     result.error.issues.forEach(issue => {
       // カスタム警告の判定
       if (issue.message.includes('未来すぎ') || issue.message.includes('古すぎ')) {
@@ -135,7 +147,7 @@ export const validateBook = (bookData: Partial<NewBook>): BookValidationResult =
         errors.push(issue.message);
       }
     });
-    
+
     return {
       isValid: errors.length === 0,
       errors,
@@ -153,7 +165,7 @@ export const validateBook = (bookData: Partial<NewBook>): BookValidationResult =
 export const validateSeries = (seriesData: Partial<NewSeries>): SeriesValidationResult => {
   try {
     const result = SERIES_VALIDATION_SCHEMA.safeParse(seriesData);
-    
+
     if (result.success) {
       return {
         isValid: true,
@@ -161,10 +173,10 @@ export const validateSeries = (seriesData: Partial<NewSeries>): SeriesValidation
         warnings: []
       };
     }
-    
+
     const errors: string[] = [];
     const warnings: string[] = [];
-    
+
     result.error.issues.forEach(issue => {
       // URL形式の警告を分類
       if (issue.path.includes('thumbnail') && issue.message.includes('URL')) {
@@ -173,7 +185,7 @@ export const validateSeries = (seriesData: Partial<NewSeries>): SeriesValidation
         errors.push(issue.message);
       }
     });
-    
+
     return {
       isValid: errors.length === 0,
       errors,
